pyproject.toml
[project]
name = "cortex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docker>=7.1.0",
    "fastmcp>=2.14.0",
    "google-adk>=1.22.1",
    "google-genai>=1.57.0",
    "ipython>=9.9.0",
    "litellm>=1.80.16",
    "mcp>=1.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sse-starlette>=2.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.15.1",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.per-file-ignores]
"api.py" = ["E402"]
"example.py" = ["E402"]
"example_compare.py" = ["E402"]
"app/agents/*/prompts.py" = ["E501"]
"app/agents/replanner/prompts.py" = ["E501"]
"app/tools/plan_toolkit.py" = ["E501"]
"cortex.py" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
